pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["Aurelio <aurelio.boelter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
apache-airflow = "^2.10.3"
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
